Refactor document

+++ app/Services/DocumentService.php
@@ -3,7 +3,14 @@
 namespace App\Services;
 
 use App\Entities\Document;
+use App\Entities\DocumentVersion;
+use App\Events\Document\DocumentCreateEvent;
+use App\Events\Document\DocumentDeleteEvent;
+use App\Events\Document\DocumentReadEvent;
+use App\Events\Document\DocumentUpdateEvent;
 use App\Repositories\DocumentRepository;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Event;
 
 /**
  * Created by Codenetix team <support@codenetix.com>
@@ -14,21 +21,27 @@ class DocumentService
      * @var DocumentRepository
      */
     protected $repository;
+    /**
+     * @var DocumentVersionService
+     */
+    private $documentVersionService;
 
     /**
      * DocumentService constructor.
      * @param DocumentRepository $repository
+     * @param DocumentVersionService $documentVersionService
      */
-    public function __construct(DocumentRepository $repository)
+    public function __construct(DocumentRepository $repository, DocumentVersionService $documentVersionService)
     {
         $this->repository = $repository;
+        $this->documentVersionService = $documentVersionService;
     }
 
     /**
      * @return mixed
      */
     public function list(){
-        return $this->repository->all();
+        return $this->repository->paginate();
     }
 
     /**
@@ -36,7 +49,10 @@ class DocumentService
      * @return Document
      */
     public function find(int $id){
-        return $this->repository->find($id);
+        $document = $this->repository->find($id);
+        Event::dispatch(new DocumentReadEvent($document));
+
+        return $document;
     }
 
     /**
@@ -44,7 +60,18 @@ class DocumentService
      * @return Document
      */
     public function create(array $data){
-        return $this->repository->create($data);
+        /** @var Document $document */
+        $document = $this->repository->create($data);
+
+        $actualVersionData = $data['actualVersion'];
+        $actualVersionData['isActual'] = 1;
+        $actualVersionData['versionName'] = 1;
+        $actualVersionData['documentId'] = $document->id;
+
+        $this->documentVersionService->create($actualVersionData);
+        Event::dispatch(new DocumentCreateEvent($document));
+
+        return $document;
     }
 
     /**
@@ -53,13 +80,73 @@ class DocumentService
      * @return mixed
      */
     public function update(array $data, int $id){
-        return $this->repository->update($data, $id);
+        /** @var Document $document */
+        $document = $this->repository->update($data, $id);
+
+        if(isset($data['actualVersionId'])) {
+            $this->setActualVersion($document->id, $data['actualVersionId']);
+        }
+
+        Event::dispatch(new DocumentUpdateEvent($document));
+
+        return $document;
+    }
+
+    /**
+     * @param array $data
+     * @param int $id
+     * @return mixed
+     */
+    public function updateVersion(array $data, int $id){
+        $createNewVersion = $data['createNewVersion'];
+
+        $document = $this->find($id);
+
+        $oldActualVersion = $document->documentActualVersion;
+        $actualVersionData = $data['actualVersion'];
+        $actualVersionData['isActual'] = 1;
+        $actualVersionData['documentId'] = $document->id;
+        $actualVersionData['versionName'] = (int)$oldActualVersion->versionName + ($createNewVersion ? 1 : 0);
+
+
+        $this->documentVersionService->create($actualVersionData);
+
+        if($createNewVersion) {
+            $this->documentVersionService->update(['isActual' => false], $oldActualVersion->id);
+        } else {
+            $this->documentVersionService->delete($oldActualVersion->id);
+        }
+
+        return $this->update($data, $id);
     }
 
     /**
      * @param int $id
      */
     public function delete(int $id){
+        $document = $this->repository->findWhere([['id', '=', $id]])->first();
+
+        if (is_null($document)) {
+            return;
+        }
+
+        Event::dispatch(new DocumentDeleteEvent($document));
+
         $this->repository->delete($id);
     }
+
+    public function setActualVersion($documentId, $versionId) {
+        $document = $this->find($documentId);
+        $newVersion = $this->documentVersionService->find($versionId);
+
+        if ($newVersion->documentId != $documentId) {
+            throw (new ModelNotFoundException())->setModel(DocumentVersion::class);
+        }
+
+        $oldVersion = $document->documentActualVersion;
+
+        $this->documentVersionService->update(['isActual' => false], $oldVersion->id);
+        $this->documentVersionService->update(['isActual' => true], $newVersion->id);
+    }
+
 }
